% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_config.R
\name{api_config}
\alias{api_config}
\title{api_config Class}
\description{
api_config Class

api_config Class
}
\details{
A R6 class to manage config.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-api_config-new}{\code{api_config$new()}}
\item \href{#method-api_config-set_api_key}{\code{api_config$set_api_key()}}
\item \href{#method-api_config-set_proxy}{\code{api_config$set_proxy()}}
\item \href{#method-api_config-get_proxy}{\code{api_config$get_proxy()}}
\item \href{#method-api_config-get_api_key}{\code{api_config$get_api_key()}}
\item \href{#method-api_config-get_api_endpoints}{\code{api_config$get_api_endpoints()}}
\item \href{#method-api_config-clone}{\code{api_config$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-api_config-new"></a>}}
\if{latex}{\out{\hypertarget{method-api_config-new}{}}}
\subsection{Method \code{new()}}{
Initialize the api_config object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{api_config$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-api_config-set_api_key"></a>}}
\if{latex}{\out{\hypertarget{method-api_config-set_api_key}{}}}
\subsection{Method \code{set_api_key()}}{
Configure the api_key settings.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{api_config$set_api_key(api_key)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{api_key}}{your openai_key}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-api_config-set_proxy"></a>}}
\if{latex}{\out{\hypertarget{method-api_config-set_proxy}{}}}
\subsection{Method \code{set_proxy()}}{
Configure the proxy settings.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{api_config$set_proxy(proxy_ip, proxy_port)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{proxy_ip}}{character Required. The IP address of the proxy.}

\item{\code{proxy_port}}{character Required. The port number of the proxy.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-api_config-get_proxy"></a>}}
\if{latex}{\out{\hypertarget{method-api_config-get_proxy}{}}}
\subsection{Method \code{get_proxy()}}{
Configure the proxy gettings.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{api_config$get_proxy()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-api_config-get_api_key"></a>}}
\if{latex}{\out{\hypertarget{method-api_config-get_api_key}{}}}
\subsection{Method \code{get_api_key()}}{
Api key gettings.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{api_config$get_api_key()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-api_config-get_api_endpoints"></a>}}
\if{latex}{\out{\hypertarget{method-api_config-get_api_endpoints}{}}}
\subsection{Method \code{get_api_endpoints()}}{
Endpoints gettings.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{api_config$get_api_endpoints()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-api_config-clone"></a>}}
\if{latex}{\out{\hypertarget{method-api_config-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{api_config$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
