% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embeddings.R
\name{embeddings}
\alias{embeddings}
\title{embeddings Class}
\description{
embeddings Class

embeddings Class
}
\section{Super class}{
\code{\link[openaistream:openai]{openaistream::openai}} -> \code{embeddings}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-embeddings-create}{\code{embeddings$create()}}
\item \href{#method-embeddings-clone}{\code{embeddings$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="openaistream" data-topic="openai" data-id="initialize"><a href='../../openaistream/html/openai.html#method-openai-initialize'><code>openaistream::openai$initialize()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-embeddings-create"></a>}}
\if{latex}{\out{\hypertarget{method-embeddings-create}{}}}
\subsection{Method \code{create()}}{
Creates an embedding vector representing the input text.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{embeddings$create(model, input, ..., verbosity = 0)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{model}}{character Required. ID of the model to use. You can use
the List models API to see all of your available models,
or see our Model overview for descriptions of them.}

\item{\code{input}}{character Required. Input text to embed, encoded as a string or array of tokens.
To embed multiple inputs in a single request, pass an array of strings or array of token arrays.
The input must not exceed the max input tokens for
the model (8192 tokens for text-embedding-ada-002), cannot be an empty string,
and any array must be 2048 dimensions or less.}

\item{\code{...}}{Additional parameters as required by the OpenAI API.For example:encoding_format;user....}

\item{\code{verbosity}}{numeric. Verbosity level for the API call(0:no output;1:show headers;
2:show headers and bodies;3: show headers, bodies, and curl status messages.).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Embeddings for the input data.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-embeddings-clone"></a>}}
\if{latex}{\out{\hypertarget{method-embeddings-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{embeddings$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
