% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openai.R
\name{openai}
\alias{openai}
\title{An R6 Class Interface to OpenAI API}
\description{
Provides methods to interact with OpenAI API including
  fetching model details, generating completions, managing files, and more.
  Always ensure that the API key is kept private.
}
\examples{
if(file.exists("openai-key")){
  key <- readLines(file("openai-key"))
  #NOTE To retrieve the correct data, it is necessary to provide an 'openaikey';
  # otherwise, an exception will be returned.
}else{
  key <- "you_api_key"
}
aaa <- openai$new(key)
#if need proxy
#aaa$set_proxy("127.0.0.1", 10890)
# List model
aaa$get_model_list()
\donttest{
#The following request will cause a timeout when the correct key is not present,
# so it is not necessary to test it during the check.
# Upload a file
train_file_path<-system.file("exdata","train.jsonl", package = "openaistream")
file_id <- aaa$files_upload(path = train_file_path,purpose = "fine-tune")
# Retrieve a file's details
aaa$files_retrieve(NULL, verbosity = 0)
# Delete a file
#aaa$files_delete(file_id$id, verbosity = 0)
# Get a list of files
aaa$files_get_list()
#aaa
job<-aaa$fine_tuning_jobs_create(model = "gpt-3.5-turbo",training_file = file_id$id)
aaa$files_delete(file_id$id, verbosity = 0)
# List fine-tuning jobs
aaa$fine_tuning_jobs_list(limit = 2, verbosity = 3)
# Retrieve details of a specific fine-tuning job
aaa$fine_tuning_jobs_retrieve(job$id, verbosity = 0)
# Get events related to a specific fine-tuning job
aaa$fine_tuning_jobs_events(job$id, verbosity = 3)
# Compute embeddings for input data
aaa$embeddings(model = "text-embedding-ada-002", input = "who are you?")
# chat test not stream
streamlg <- aaa$get_completions_query(
  prompt = "Please explain the World Wars?",
  model = "davinci-002",
  stream = FALSE,
  max_tokens = 20,
  num = 4,
  verbosity = 0
)

# chat test stream
streamlg <- aaa$get_chat_completions_query(
  messages = data.frame(role = c("system", "user"),
                        content = c("You are a assistant.", "How's the weather today?")),
  model = "gpt-3.5-turbo",
  stream = TRUE,
  max_tokens = 10
)

streamlg$get_state()
streamlg$next_value
streamlg$close()
streamlg$next_value
streamlg$get_state()
sss<-aaa$audio_speech(input = "Hi, this is a voice transmission test.",response_format="mp3")
#writeBin(sss,"test.mp3")

streammp3<-aaa$audio_speech(input = "Hi, this is a voice transmission test",
              response_format="mp3",stream = TRUE,num=1000)
streammp3$get_state()
#fcon <- file("yourfile.mp3", "wb")
#for(i in 1:3){
#   writeBin(streammp3$next_value,fcon)
#}
streammp3$close()
streammp3$next_value
streammp3$get_state()

#text_E1<-aaa$audio_transcriptions(path = "test.mp3")
#text_E2<-aaa$audio_translations(path = "test.mp3")
pic1<-aaa$images_generations(prompt = "A small bird flies over the ocean")
#req <- request(pic1$data$url) \%>\%req_perform(path = "pic1.png")
#Here is the image test.
#For simple test,you need install 'png' package.
#Note that when a mask is not provided, OpenAI defaults to using areas of the uploaded image
#  with a transparency of 0 as the modification area.
#The following test adds full image transparency of 0 (modifying the entire image) to pic1.
# pic1<-readPNG("pic1.png")
# if (dim(pic1)[3] == 3) {  # Checking if the image is RGB
#   # Adding an alpha channel (fully opaque)
#   alpha_channel <- matrix(0, nrow = dim(pic1)[1], ncol = dim(pic1)[2])
#   pic1_b<-array(c(pic1,alpha_channel), c(1024,1024,4))
# }
# writePNG(pic1_b,target = "pic1_b.png")
# pic3<-aaa$images_edits(image = "pic1_b.png",
# prompt = "Please modify the bird in the picture to a flock of birds."
#  ,verbosity = 3,n="2",size="256x256")
# request(pic3$data$url[1]) \%>\%req_perform(path = "pic3_1.png")
# request(pic3$data$url[2]) \%>\%req_perform(path = "pic3_2.png")
#
# pic4<-aaa$images_variations(image = "pic1.png",n="2",size="256x256")
# request(pic4$data$url[1]) \%>\%req_perform(path = "pic4_1.png")
# request(pic4$data$url[2]) \%>\%req_perform(path = "pic4_2.png")
}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-openai-new}{\code{openai$new()}}
\item \href{#method-openai-set_proxy}{\code{openai$set_proxy()}}
\item \href{#method-openai-get_model_list}{\code{openai$get_model_list()}}
\item \href{#method-openai-get_model_retrieve}{\code{openai$get_model_retrieve()}}
\item \href{#method-openai-fine_tuning_jobs_delete}{\code{openai$fine_tuning_jobs_delete()}}
\item \href{#method-openai-get_completions_query}{\code{openai$get_completions_query()}}
\item \href{#method-openai-get_chat_completions_query}{\code{openai$get_chat_completions_query()}}
\item \href{#method-openai-files_upload}{\code{openai$files_upload()}}
\item \href{#method-openai-files_get_list}{\code{openai$files_get_list()}}
\item \href{#method-openai-files_delete}{\code{openai$files_delete()}}
\item \href{#method-openai-files_retrieve}{\code{openai$files_retrieve()}}
\item \href{#method-openai-fine_tuning_jobs_create}{\code{openai$fine_tuning_jobs_create()}}
\item \href{#method-openai-fine_tuning_jobs_list}{\code{openai$fine_tuning_jobs_list()}}
\item \href{#method-openai-fine_tuning_jobs_retrieve}{\code{openai$fine_tuning_jobs_retrieve()}}
\item \href{#method-openai-fine_tuning_jobs_cancel}{\code{openai$fine_tuning_jobs_cancel()}}
\item \href{#method-openai-fine_tuning_jobs_events}{\code{openai$fine_tuning_jobs_events()}}
\item \href{#method-openai-embeddings}{\code{openai$embeddings()}}
\item \href{#method-openai-audio_speech}{\code{openai$audio_speech()}}
\item \href{#method-openai-audio_transcriptions}{\code{openai$audio_transcriptions()}}
\item \href{#method-openai-audio_translations}{\code{openai$audio_translations()}}
\item \href{#method-openai-images_generations}{\code{openai$images_generations()}}
\item \href{#method-openai-images_edits}{\code{openai$images_edits()}}
\item \href{#method-openai-images_variations}{\code{openai$images_variations()}}
\item \href{#method-openai-assistants_create}{\code{openai$assistants_create()}}
\item \href{#method-openai-assistants_retrieve}{\code{openai$assistants_retrieve()}}
\item \href{#method-openai-assistants_modify}{\code{openai$assistants_modify()}}
\item \href{#method-openai-assistants_delete}{\code{openai$assistants_delete()}}
\item \href{#method-openai-assistants_list}{\code{openai$assistants_list()}}
\item \href{#method-openai-assistants_file_create}{\code{openai$assistants_file_create()}}
\item \href{#method-openai-assistants_file_retrieve}{\code{openai$assistants_file_retrieve()}}
\item \href{#method-openai-assistants_file_delete}{\code{openai$assistants_file_delete()}}
\item \href{#method-openai-assistants_file_list}{\code{openai$assistants_file_list()}}
\item \href{#method-openai-threads_create}{\code{openai$threads_create()}}
\item \href{#method-openai-threads_retrieve}{\code{openai$threads_retrieve()}}
\item \href{#method-openai-threads_modify}{\code{openai$threads_modify()}}
\item \href{#method-openai-threads_delete}{\code{openai$threads_delete()}}
\item \href{#method-openai-messages_create}{\code{openai$messages_create()}}
\item \href{#method-openai-messages_retrieve}{\code{openai$messages_retrieve()}}
\item \href{#method-openai-messages_modify}{\code{openai$messages_modify()}}
\item \href{#method-openai-messages_list}{\code{openai$messages_list()}}
\item \href{#method-openai-messages_file_retrieve}{\code{openai$messages_file_retrieve()}}
\item \href{#method-openai-messages_file_list}{\code{openai$messages_file_list()}}
\item \href{#method-openai-runs_create}{\code{openai$runs_create()}}
\item \href{#method-openai-runs_retrieve}{\code{openai$runs_retrieve()}}
\item \href{#method-openai-runs_modify}{\code{openai$runs_modify()}}
\item \href{#method-openai-runs_list}{\code{openai$runs_list()}}
\item \href{#method-openai-runs_submit_tool_outputs}{\code{openai$runs_submit_tool_outputs()}}
\item \href{#method-openai-runs_cancel}{\code{openai$runs_cancel()}}
\item \href{#method-openai-runs_create_tread}{\code{openai$runs_create_tread()}}
\item \href{#method-openai-runs_steps_retrieve}{\code{openai$runs_steps_retrieve()}}
\item \href{#method-openai-runs_steps_list}{\code{openai$runs_steps_list()}}
\item \href{#method-openai-clone}{\code{openai$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-openai-new"></a>}}
\if{latex}{\out{\hypertarget{method-openai-new}{}}}
\subsection{Method \code{new()}}{
Initialize the OpenAI API interface with the provided API key.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{openai$new(api_key)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{api_key}}{The OpenAI API key.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-openai-set_proxy"></a>}}
\if{latex}{\out{\hypertarget{method-openai-set_proxy}{}}}
\subsection{Method \code{set_proxy()}}{
Configure the proxy settings.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{openai$set_proxy(proxy_ip, proxy_port)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{proxy_ip}}{character Required. The IP address of the proxy.}

\item{\code{proxy_port}}{character Required. The port number of the proxy.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-openai-get_model_list"></a>}}
\if{latex}{\out{\hypertarget{method-openai-get_model_list}{}}}
\subsection{Method \code{get_model_list()}}{
Lists the currently available models, and provides basic information
             about each one such as the owner and availability.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{openai$get_model_list(verbosity = 0)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{verbosity}}{numeric. Verbosity level for the API call(0:no output;1:show headers;
2:show headers and bodies;3: show headers, bodies, and curl status messages.).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list of model objects.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-openai-get_model_retrieve"></a>}}
\if{latex}{\out{\hypertarget{method-openai-get_model_retrieve}{}}}
\subsection{Method \code{get_model_retrieve()}}{
Retrieves a model instance, providing basic information about
             the model such as the owner and permissioning.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{openai$get_model_retrieve(model, verbosity = 0)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{model}}{character Required. The ID of the model to use for this request}

\item{\code{verbosity}}{numeric. Verbosity level for the API call(0:no output;1:show headers;
2:show headers and bodies;3: show headers, bodies, and curl status messages.).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The model object matching the specified ID.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-openai-fine_tuning_jobs_delete"></a>}}
\if{latex}{\out{\hypertarget{method-openai-fine_tuning_jobs_delete}{}}}
\subsection{Method \code{fine_tuning_jobs_delete()}}{
Delete a fine-tuned model. You must have the Owner role in
             your organization to delete a model.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{openai$fine_tuning_jobs_delete(model, verbosity = 0)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{model}}{character Required. The model to delete}

\item{\code{verbosity}}{numeric. Verbosity level for the API call(0:no output;1:show headers;
2:show headers and bodies;3: show headers, bodies, and curl status messages.).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Deletion status.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-openai-get_completions_query"></a>}}
\if{latex}{\out{\hypertarget{method-openai-get_completions_query}{}}}
\subsection{Method \code{get_completions_query()}}{
Creates a completion for the provided prompt and parameters.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{openai$get_completions_query(
  prompt,
  model,
  stream = F,
  n = 1,
  num = 2,
  verbosity = 0,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{prompt}}{character Required. The prompt(s) to generate completions for, encoded as a string,
array of strings, array of tokens, or array of token arrays. Note that <|endoftext|> is
the document separator that the model sees during training, so if a prompt is not specified
the model will generate as if from the beginning of a new document.}

\item{\code{model}}{character Required. The model to use for generating completions.}

\item{\code{stream}}{logical. Whether to stream back partial progress. If set, tokens will be sent
as data-only server-sent events as they become available.}

\item{\code{n}}{integer. How many chat completion choices to generate for each input message. Note that you will be
charged based on the number of generated tokens across all of the choices. Keep n as 1 to minimize costs.
NOTE: The parameter sometimes fails to work when 'num' is not assigned a value, and the reason for this
is currently unclear. When failure occurs, try assigning any integer greater than 0 to the 'num' parameter.}

\item{\code{num}}{The num parameter controls the number of text entries returned by a stream in one go.
Note that this is different from the n parameter, which specifies the number of results returned.
For detailed information on the n parameter, please refer to OpenAI's API documentation.}

\item{\code{verbosity}}{numeric. Verbosity level for the API call(0:no output;1:show headers;
2:show headers and bodies;3: show headers, bodies, and curl status messages.).}

\item{\code{...}}{Additional parameters as required by the OpenAI API.For example:max_tokens;n;stop;temperature......}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Returns a completion object, or a sequence of completion objects if the request is streamed.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-openai-get_chat_completions_query"></a>}}
\if{latex}{\out{\hypertarget{method-openai-get_chat_completions_query}{}}}
\subsection{Method \code{get_chat_completions_query()}}{
Creates a model response for the given chat conversation.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{openai$get_chat_completions_query(
  messages,
  model,
  stream = F,
  n = 1,
  num = 2,
  verbosity = 0,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{messages}}{list Required. A list of messages comprising the conversation so far.}

\item{\code{model}}{character Required. The model to use for generating chat completions.}

\item{\code{stream}}{logical. Whether to stream back partial progress. If set, tokens will be sent
as data-only server-sent events as they become available.}

\item{\code{n}}{integer. How many chat completion choices to generate for each input message. Note that you will be
charged based on the number of generated tokens across all of the choices. Keep n as 1 to minimize costs.
NOTE: The parameter sometimes fails to work when 'num' is not assigned a value, and the reason for this
is currently unclear. When failure occurs, try assigning any integer greater than 0 to the 'num' parameter.}

\item{\code{num}}{The num parameter controls the number of text entries returned by a stream in one go.
Note that this is different from the n parameter, which specifies the number of results returned.
For detailed information on the n parameter, please refer to OpenAI's API documentation.}

\item{\code{verbosity}}{numeric. Verbosity level for the API call(0:no output;1:show headers;
2:show headers and bodies;3: show headers, bodies, and curl status messages.).}

\item{\code{...}}{Additional parameters as required by the OpenAI API.For example:max_tokens;n;stop;temperature......}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Returns a chat completion object, or a streamed sequence of chat completion chunk objects if the request is streamed.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-openai-files_upload"></a>}}
\if{latex}{\out{\hypertarget{method-openai-files_upload}{}}}
\subsection{Method \code{files_upload()}}{
Upload a file that can be used across various endpoints. The size of all the
             files uploaded by one organization can be up to 100 GB.The size of individual
             files can be a maximum of 512 MB or 2 million tokens for Assistants.
             See the Assistants Tools guide to learn more about the types of files supported.
             The Fine-tuning API only supports .jsonl files.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{openai$files_upload(path, verbosity = 0, purpose)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{character Required. Path to the file that needs to be uploaded.}

\item{\code{verbosity}}{numeric. Verbosity level for the API call(0:no output;1:show headers;
2:show headers and bodies;3: show headers, bodies, and curl status messages.).}

\item{\code{purpose}}{The intended purpose of the uploaded file.
Use "fine-tune" for Fine-tuning and "assistants" for Assistants and Messages.
This allows us to validate the format of the uploaded file is correct for fine-tuning.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The uploaded File object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-openai-files_get_list"></a>}}
\if{latex}{\out{\hypertarget{method-openai-files_get_list}{}}}
\subsection{Method \code{files_get_list()}}{
Returns a list of files that belong to the user's organization.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{openai$files_get_list(..., verbosity = 0)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Additional parameters as required by the OpenAI API.For example:purpose}

\item{\code{verbosity}}{numeric. Verbosity level for the API call(0:no output;1:show headers;
2:show headers and bodies;3: show headers, bodies, and curl status messages.).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list of File objects.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-openai-files_delete"></a>}}
\if{latex}{\out{\hypertarget{method-openai-files_delete}{}}}
\subsection{Method \code{files_delete()}}{
Delete a file.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{openai$files_delete(file_id, verbosity = 0)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{file_id}}{character Required. The ID of the file to use for this request.}

\item{\code{verbosity}}{numeric. Verbosity level for the API call(0:no output;1:show headers;
2:show headers and bodies;3: show headers, bodies, and curl status messages.).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Deletion status.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-openai-files_retrieve"></a>}}
\if{latex}{\out{\hypertarget{method-openai-files_retrieve}{}}}
\subsection{Method \code{files_retrieve()}}{
Returns information about a specific file.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{openai$files_retrieve(file_id, verbosity = 0)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{file_id}}{character Required. The ID of the file to retrieve details for.}

\item{\code{verbosity}}{numeric. Verbosity level for the API call(0:no output;1:show headers;
2:show headers and bodies;3: show headers, bodies, and curl status messages.).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The File object matching the specified ID.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-openai-fine_tuning_jobs_create"></a>}}
\if{latex}{\out{\hypertarget{method-openai-fine_tuning_jobs_create}{}}}
\subsection{Method \code{fine_tuning_jobs_create()}}{
Creates a job that fine-tunes a specified model from a given dataset.
             Response includes details of the enqueued job including job status and
             the name of the fine-tuned models once complete.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{openai$fine_tuning_jobs_create(
  model,
  training_file,
  hyperparameters = list(n_epochs = 1),
  ...,
  verbosity = 0
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{model}}{character Required. The model ID.}

\item{\code{training_file}}{character Required. The file used for training.}

\item{\code{hyperparameters}}{list. The hyperparameters used for the fine-tuning job. include batch_size;learning_rate_multiplier;n_epochs.}

\item{\code{...}}{Additional parameters as required by the OpenAI API.For example:validation_file......}

\item{\code{verbosity}}{numeric. Verbosity level for the API call(0:no output;1:show headers;
2:show headers and bodies;3: show headers, bodies, and curl status messages.).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Response indicating the success or failure of the fine-tuning job creation.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-openai-fine_tuning_jobs_list"></a>}}
\if{latex}{\out{\hypertarget{method-openai-fine_tuning_jobs_list}{}}}
\subsection{Method \code{fine_tuning_jobs_list()}}{
List your organization's fine-tuning jobs
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{openai$fine_tuning_jobs_list(verbosity = 0, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{verbosity}}{numeric. Verbosity level for the API call(0:no output;1:show headers;
2:show headers and bodies;3: show headers, bodies, and curl status messages.).}

\item{\code{...}}{Additional parameters as required by the OpenAI API. For example:after,limit.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list of paginated fine-tuning job objects.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-openai-fine_tuning_jobs_retrieve"></a>}}
\if{latex}{\out{\hypertarget{method-openai-fine_tuning_jobs_retrieve}{}}}
\subsection{Method \code{fine_tuning_jobs_retrieve()}}{
Get info about a fine-tuning job.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{openai$fine_tuning_jobs_retrieve(job_id, verbosity = 0)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{job_id}}{character Required. The ID of the fine-tuning job.}

\item{\code{verbosity}}{numeric. Verbosity level for the API call(0:no output;1:show headers;
2:show headers and bodies;3: show headers, bodies, and curl status messages.).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The fine-tuning object with the given ID.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-openai-fine_tuning_jobs_cancel"></a>}}
\if{latex}{\out{\hypertarget{method-openai-fine_tuning_jobs_cancel}{}}}
\subsection{Method \code{fine_tuning_jobs_cancel()}}{
Immediately cancel a fine-tune job.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{openai$fine_tuning_jobs_cancel(job_id, verbosity = 0)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{job_id}}{character Required. The ID of the fine-tuning job to cancel.}

\item{\code{verbosity}}{numeric. Verbosity level for the API call(0:no output;1:show headers;
2:show headers and bodies;3: show headers, bodies, and curl status messages.).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The cancelled fine-tuning object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-openai-fine_tuning_jobs_events"></a>}}
\if{latex}{\out{\hypertarget{method-openai-fine_tuning_jobs_events}{}}}
\subsection{Method \code{fine_tuning_jobs_events()}}{
Get status updates for a fine-tuning job.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{openai$fine_tuning_jobs_events(job_id, ..., verbosity = 0)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{job_id}}{character Required. The ID of the fine-tuning job to get events for.}

\item{\code{...}}{Additional parameters as required by the OpenAI API. For example:after,limit.}

\item{\code{verbosity}}{numeric. Verbosity level for the API call(0:no output;1:show headers;
2:show headers and bodies;3: show headers, bodies, and curl status messages.).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list of fine-tuning event objects.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-openai-embeddings"></a>}}
\if{latex}{\out{\hypertarget{method-openai-embeddings}{}}}
\subsection{Method \code{embeddings()}}{
Creates an embedding vector representing the input text.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{openai$embeddings(model, input, ..., verbosity = 0)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{model}}{character Required. ID of the model to use. You can use
the List models API to see all of your available models,
or see our Model overview for descriptions of them.}

\item{\code{input}}{character Required. Input text to embed, encoded as a string or array of tokens.
To embed multiple inputs in a single request, pass an array of strings or array of token arrays.
The input must not exceed the max input tokens for
the model (8192 tokens for text-embedding-ada-002), cannot be an empty string,
and any array must be 2048 dimensions or less.}

\item{\code{...}}{Additional parameters as required by the OpenAI API.For example:encoding_format;user....}

\item{\code{verbosity}}{numeric. Verbosity level for the API call(0:no output;1:show headers;
2:show headers and bodies;3: show headers, bodies, and curl status messages.).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Embeddings for the input data.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-openai-audio_speech"></a>}}
\if{latex}{\out{\hypertarget{method-openai-audio_speech}{}}}
\subsection{Method \code{audio_speech()}}{
Generates audio from the input text.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{openai$audio_speech(
  model = "tts-1",
  input,
  voice = "alloy",
  stream = F,
  num = 100,
  ...,
  verbosity = 0
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{model}}{character Required. One of the available TTS models: tts-1 or tts-1-hd}

\item{\code{input}}{character Required. The text to generate audio for. The maximum length is 4096 characters.}

\item{\code{voice}}{character Required. The voice to use when generating the audio. Supported voices are alloy, echo, fable, onyx, nova, and shimmer.}

\item{\code{stream}}{logical. Using the stream call, it will return raw data of the specified length,
which can be saved in the set format such as mp3, etc. For details, please see the examples.}

\item{\code{num}}{The num parameter controls the number of raw entries returned by a stream in one go.
Note that this is different from the n parameter, which specifies the number of results returned.
For detailed information on the n parameter, please refer to OpenAI's API documentation.}

\item{\code{...}}{Additional parameters as required by the OpenAI API.For example:response_format;speed....}

\item{\code{verbosity}}{numeric. Verbosity level for the API call(0:no output;1:show headers;
2:show headers and bodies;3: show headers, bodies, and curl status messages.).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The audio file content.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-openai-audio_transcriptions"></a>}}
\if{latex}{\out{\hypertarget{method-openai-audio_transcriptions}{}}}
\subsection{Method \code{audio_transcriptions()}}{
Transcribes audio into the input language.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{openai$audio_transcriptions(path, model = "whisper-1", ..., verbosity = 0)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{character Required. The audio file object (not file name) to transcribe,
in one of these formats: flac, mp3, mp4, mpeg, mpga, m4a, ogg, wav, or webm.}

\item{\code{model}}{character Required. ID of the model to use. Only whisper-1 is currently available.}

\item{\code{...}}{Additional parameters as required by the OpenAI API.For example:language;prompt;response_format;temperature....}

\item{\code{verbosity}}{numeric. Verbosity level for the API call(0:no output;1:show headers;
2:show headers and bodies;3: show headers, bodies, and curl status messages.).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The transcribed text.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-openai-audio_translations"></a>}}
\if{latex}{\out{\hypertarget{method-openai-audio_translations}{}}}
\subsection{Method \code{audio_translations()}}{
Translates audio into English.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{openai$audio_translations(path, model = "whisper-1", ..., verbosity = 0)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{character Required. The audio file object (not file name) to transcribe,
in one of these formats: flac, mp3, mp4, mpeg, mpga, m4a, ogg, wav, or webm.}

\item{\code{model}}{character Required. ID of the model to use. Only whisper-1 is currently available.}

\item{\code{...}}{Additional parameters as required by the OpenAI API.For example:prompt;response_format;temperature....}

\item{\code{verbosity}}{numeric. Verbosity level for the API call(0:no output;1:show headers;
2:show headers and bodies;3: show headers, bodies, and curl status messages.).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The transcribed text.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-openai-images_generations"></a>}}
\if{latex}{\out{\hypertarget{method-openai-images_generations}{}}}
\subsection{Method \code{images_generations()}}{
Creates an image given a prompt.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{openai$images_generations(prompt, ..., verbosity = 0)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{prompt}}{character Required. A text description of the desired image(s).
The maximum length is 1000 characters for dall-e-2 and 4000 characters for dall-e-3}

\item{\code{...}}{Additional parameters as required by the OpenAI API.For example:n;quality;response_format...}

\item{\code{verbosity}}{numeric. Verbosity level for the API call(0:no output;1:show headers;
2:show headers and bodies;3: show headers, bodies, and curl status messages.).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Returns a list of image objects.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-openai-images_edits"></a>}}
\if{latex}{\out{\hypertarget{method-openai-images_edits}{}}}
\subsection{Method \code{images_edits()}}{
Creates an edited or extended image given an original image and a prompt.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{openai$images_edits(image, prompt, ..., verbosity = 0)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{image}}{character Required. The image to edit. Must be a valid PNG file, less than 4MB,
and square. If mask is not provided, image must have transparency, which will be used as the mask.}

\item{\code{prompt}}{character Required. A text description of the desired image(s). The maximum length is 1000 characters.}

\item{\code{...}}{Additional parameters as required by the OpenAI API.For example:mask;model;n...}

\item{\code{verbosity}}{numeric. Verbosity level for the API call(0:no output;1:show headers;
2:show headers and bodies;3: show headers, bodies, and curl status messages.).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Returns a list of image objects.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-openai-images_variations"></a>}}
\if{latex}{\out{\hypertarget{method-openai-images_variations}{}}}
\subsection{Method \code{images_variations()}}{
Creates a variation of a given image.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{openai$images_variations(image, ..., verbosity = 0)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{image}}{character Required. The image to use as the basis for the variation(s). Must be a valid PNG file, less than 4MB, and square.}

\item{\code{...}}{Additional parameters as required by the OpenAI API.For example:model;n;response_format}

\item{\code{verbosity}}{numeric. Verbosity level for the API call(0:no output;1:show headers;
2:show headers and bodies;3: show headers, bodies, and curl status messages.).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Returns a list of image objects.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-openai-assistants_create"></a>}}
\if{latex}{\out{\hypertarget{method-openai-assistants_create}{}}}
\subsection{Method \code{assistants_create()}}{
Create an assistant with a model and instructions.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{openai$assistants_create(model, ..., verbosity = 0)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{model}}{character Required. ID of the model to use. You can use the List models API to see all of your available models,
or see our Model overview for descriptions of them.}

\item{\code{...}}{Additional parameters as required by the OpenAI API.For example:name;description;instructions;tools;file_ids;metadata}

\item{\code{verbosity}}{numeric. Verbosity level for the API call(0:no output;1:show headers;
2:show headers and bodies;3: show headers, bodies, and curl status messages.).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An assistant object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-openai-assistants_retrieve"></a>}}
\if{latex}{\out{\hypertarget{method-openai-assistants_retrieve}{}}}
\subsection{Method \code{assistants_retrieve()}}{
Retrieves an assistant.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{openai$assistants_retrieve(assistant_id, verbosity = 0)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{assistant_id}}{character Required. The ID of the assistant to retrieve.}

\item{\code{verbosity}}{numeric. Verbosity level for the API call(0:no output;1:show headers;
2:show headers and bodies;3: show headers, bodies, and curl status messages.).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The assistant object matching the specified ID.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-openai-assistants_modify"></a>}}
\if{latex}{\out{\hypertarget{method-openai-assistants_modify}{}}}
\subsection{Method \code{assistants_modify()}}{
Modifies an assistant.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{openai$assistants_modify(assistant_id, ..., verbosity = 0)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{assistant_id}}{character Required. The ID of the assistant to modify.}

\item{\code{...}}{Additional parameters as required by the OpenAI API.For example:model;name;description;instructions;tools;file_ids,metadata}

\item{\code{verbosity}}{numeric. Verbosity level for the API call(0:no output;1:show headers;
2:show headers and bodies;3: show headers, bodies, and curl status messages.).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The assistant object matching the specified ID.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-openai-assistants_delete"></a>}}
\if{latex}{\out{\hypertarget{method-openai-assistants_delete}{}}}
\subsection{Method \code{assistants_delete()}}{
Delete an assistant.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{openai$assistants_delete(assistant_id, verbosity = 0)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{assistant_id}}{character Required. The ID of the assistant to delete.}

\item{\code{verbosity}}{numeric. Verbosity level for the API call(0:no output;1:show headers;
2:show headers and bodies;3: show headers, bodies, and curl status messages.).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Deletion status
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-openai-assistants_list"></a>}}
\if{latex}{\out{\hypertarget{method-openai-assistants_list}{}}}
\subsection{Method \code{assistants_list()}}{
Returns a list of assistants
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{openai$assistants_list(..., verbosity = 0)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Additional parameters as required by the OpenAI API.For example:limit;order;after;before;}

\item{\code{verbosity}}{numeric. Verbosity level for the API call(0:no output;1:show headers;
2:show headers and bodies;3: show headers, bodies, and curl status messages.).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list of assistant objects.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-openai-assistants_file_create"></a>}}
\if{latex}{\out{\hypertarget{method-openai-assistants_file_create}{}}}
\subsection{Method \code{assistants_file_create()}}{
Create an assistant file by attaching a File to an assistant.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{openai$assistants_file_create(assistant_id, file_id, verbosity = 0)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{assistant_id}}{character Required. The ID of the assistant for which to create a File.}

\item{\code{file_id}}{character Required. A File ID (with purpose="assistants") that the assistant should use.
Useful for tools like retrieval and code_interpreter that can access files.}

\item{\code{verbosity}}{numeric. Verbosity level for the API call(0:no output;1:show headers;
2:show headers and bodies;3: show headers, bodies, and curl status messages.).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An assistant file object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-openai-assistants_file_retrieve"></a>}}
\if{latex}{\out{\hypertarget{method-openai-assistants_file_retrieve}{}}}
\subsection{Method \code{assistants_file_retrieve()}}{
Retrieves an AssistantFile.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{openai$assistants_file_retrieve(assistant_id, file_id, verbosity = 0)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{assistant_id}}{character Required. The ID of the assistant who the file belongs to.}

\item{\code{file_id}}{character Required. The ID of the file we're getting.}

\item{\code{verbosity}}{numeric. Verbosity level for the API call(0:no output;1:show headers;
2:show headers and bodies;3: show headers, bodies, and curl status messages.).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The assistant file object matching the specified ID.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-openai-assistants_file_delete"></a>}}
\if{latex}{\out{\hypertarget{method-openai-assistants_file_delete}{}}}
\subsection{Method \code{assistants_file_delete()}}{
Delete an assistant file.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{openai$assistants_file_delete(assistant_id, file_id, verbosity = 0)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{assistant_id}}{character Required. The ID of the assistant who the file belongs to.}

\item{\code{file_id}}{character Required. The ID of the file we're getting.}

\item{\code{verbosity}}{numeric. Verbosity level for the API call(0:no output;1:show headers;
2:show headers and bodies;3: show headers, bodies, and curl status messages.).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Deletion status
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-openai-assistants_file_list"></a>}}
\if{latex}{\out{\hypertarget{method-openai-assistants_file_list}{}}}
\subsection{Method \code{assistants_file_list()}}{
Retrieve a list of files from OpenAI.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{openai$assistants_file_list(assistant_id, ..., verbosity = 0)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{assistant_id}}{character Required. The ID of the assistant who the file belongs to.}

\item{\code{...}}{Additional parameters as required by the OpenAI API.For example:limit,order,after,before}

\item{\code{verbosity}}{numeric. Verbosity level for the API call(0:no output;1:show headers;
2:show headers and bodies;3: show headers, bodies, and curl status messages.).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list of files.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-openai-threads_create"></a>}}
\if{latex}{\out{\hypertarget{method-openai-threads_create}{}}}
\subsection{Method \code{threads_create()}}{
Create a thread.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{openai$threads_create(..., verbosity = 0)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Additional parameters as required by the OpenAI API.For example:messages;name;metadata}

\item{\code{verbosity}}{numeric Verbosity level for the API call(0:no output;1:show headers;
2:show headers and bodies;3: show headers, bodies, and curl status messages.)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A thread object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-openai-threads_retrieve"></a>}}
\if{latex}{\out{\hypertarget{method-openai-threads_retrieve}{}}}
\subsection{Method \code{threads_retrieve()}}{
Retrieves a thread.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{openai$threads_retrieve(thread_id, verbosity = 0)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{thread_id}}{character Required. The ID of the thread to retrieve.}

\item{\code{verbosity}}{numeric Verbosity level for the API call(0:no output;1:show headers;
2:show headers and bodies;3: show headers, bodies, and curl status messages.)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The thread object matching the specified ID.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-openai-threads_modify"></a>}}
\if{latex}{\out{\hypertarget{method-openai-threads_modify}{}}}
\subsection{Method \code{threads_modify()}}{
Modifies a thread.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{openai$threads_modify(thread_id, ..., verbosity = 0)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{thread_id}}{The ID of the thread to modify. Only the metadata can be modified.}

\item{\code{...}}{Additional parameters as required by the OpenAI API.For example:metadata}

\item{\code{verbosity}}{numeric Verbosity level for the API call(0:no output;1:show headers;
2:show headers and bodies;3: show headers, bodies, and curl status messages.)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The modified thread object matching the specified ID.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-openai-threads_delete"></a>}}
\if{latex}{\out{\hypertarget{method-openai-threads_delete}{}}}
\subsection{Method \code{threads_delete()}}{
Delete a thread.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{openai$threads_delete(thread_id, verbosity = 0)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{thread_id}}{character Required The ID of the thread to delete.}

\item{\code{verbosity}}{numeric Verbosity level for the API call(0:no output;1:show headers;
2:show headers and bodies;3: show headers, bodies, and curl status messages.)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Deletion status.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-openai-messages_create"></a>}}
\if{latex}{\out{\hypertarget{method-openai-messages_create}{}}}
\subsection{Method \code{messages_create()}}{
Create a message.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{openai$messages_create(thread_id, role, content, ..., verbosity = 0)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{thread_id}}{character Required. The ID of the thread to create a message for.}

\item{\code{role}}{character Required. The role of the entity that is creating the message. Currently only user is supported.}

\item{\code{content}}{character Required. The content of the message.}

\item{\code{...}}{Additional parameters as required by the OpenAI API. For example:file_ids,metadata}

\item{\code{verbosity}}{numeric. Verbosity level for the API call(0:no output;1:show headers;
2:show headers and bodies;3: show headers, bodies, and curl status messages.)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A message object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-openai-messages_retrieve"></a>}}
\if{latex}{\out{\hypertarget{method-openai-messages_retrieve}{}}}
\subsection{Method \code{messages_retrieve()}}{
Retrieve a message.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{openai$messages_retrieve(thread_id, message_id, verbosity = 0)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{thread_id}}{character Required. The ID of the thread the message belongs to.}

\item{\code{message_id}}{character Required. The ID of the message to retrieve.}

\item{\code{verbosity}}{numeric. Verbosity level for the API call(0:no output;1:show headers;
2:show headers and bodies;3: show headers, bodies, and curl status messages.)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The message object matching the specified ID.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-openai-messages_modify"></a>}}
\if{latex}{\out{\hypertarget{method-openai-messages_modify}{}}}
\subsection{Method \code{messages_modify()}}{
Modifies a message.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{openai$messages_modify(thread_id, message_id, ..., verbosity = 0)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{thread_id}}{character Required. The ID of the thread the message belongs to.}

\item{\code{message_id}}{character Required. The ID of the message to retrieve.}

\item{\code{...}}{Additional parameters as required by the OpenAI API. For example:metadata}

\item{\code{verbosity}}{numeric. Verbosity level for the API call(0:no output;1:show headers;
2:show headers and bodies;3: show headers, bodies, and curl status messages.)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The modified message object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-openai-messages_list"></a>}}
\if{latex}{\out{\hypertarget{method-openai-messages_list}{}}}
\subsection{Method \code{messages_list()}}{
Returns a list of messages for a given thread.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{openai$messages_list(thread_id, ..., verbosity = 0)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{thread_id}}{character Required. The ID of the thread the messages belong to.}

\item{\code{...}}{Additional parameters as required by the OpenAI API.}

\item{\code{verbosity}}{numeric Verbosity level for the API call(0:no output;1:show headers;
2:show headers and bodies;3: show headers, bodies, and curl status messages.)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list of message objects.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-openai-messages_file_retrieve"></a>}}
\if{latex}{\out{\hypertarget{method-openai-messages_file_retrieve}{}}}
\subsection{Method \code{messages_file_retrieve()}}{
Retrieves a message file.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{openai$messages_file_retrieve(thread_id, message_id, file_id, verbosity = 0)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{thread_id}}{character Required. The ID of the thread the message belongs to.}

\item{\code{message_id}}{character Required. The ID of the message the file belongs to.}

\item{\code{file_id}}{character Required The ID of the file being retrieved.}

\item{\code{verbosity}}{numeric Verbosity level for the API call(0:no output;1:show headers;}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The message file object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-openai-messages_file_list"></a>}}
\if{latex}{\out{\hypertarget{method-openai-messages_file_list}{}}}
\subsection{Method \code{messages_file_list()}}{
Returns a list of message files.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{openai$messages_file_list(thread_id, message_id, ..., verbosity = 0)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{thread_id}}{character Required. The ID of the thread the message belongs to.}

\item{\code{message_id}}{character Required. The ID of the message the file belongs to.}

\item{\code{...}}{Additional parameters as required by the OpenAI API.}

\item{\code{verbosity}}{numeric Verbosity level for the API call(0:no output;1:show headers;}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list of message file objects.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-openai-runs_create"></a>}}
\if{latex}{\out{\hypertarget{method-openai-runs_create}{}}}
\subsection{Method \code{runs_create()}}{
Create a run.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{openai$runs_create(thread_id, assistant_id, ..., verbosity = 0)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{thread_id}}{character Required. The ID of the thread to run.}

\item{\code{assistant_id}}{character Required. The ID of the assistant to use to execute this run.}

\item{\code{...}}{Additional parameters as required by the OpenAI API.For example:model,instructions,tools,metadata}

\item{\code{verbosity}}{numeric Verbosity level for the API call(0:no output;1:show headers;
2:show headers and bodies;3: show headers, bodies, and curl status messages.)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A run object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-openai-runs_retrieve"></a>}}
\if{latex}{\out{\hypertarget{method-openai-runs_retrieve}{}}}
\subsection{Method \code{runs_retrieve()}}{
Retrieves a run.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{openai$runs_retrieve(thread_id, run_id, verbosity = 0)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{thread_id}}{character Required The ID of the thread the run belongs to.}

\item{\code{run_id}}{character Required The ID of the run to retrieve.}

\item{\code{verbosity}}{numeric Verbosity level for the API call(0:no output;1:show headers;
2:show headers and bodies;3: show headers, bodies, and curl status messages.)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The run object matching the specified ID.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-openai-runs_modify"></a>}}
\if{latex}{\out{\hypertarget{method-openai-runs_modify}{}}}
\subsection{Method \code{runs_modify()}}{
Modifies a run.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{openai$runs_modify(thread_id, run_id, ..., verbosity = 0)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{thread_id}}{character Required The ID of the thread the run belongs to.}

\item{\code{run_id}}{character Required The ID of the run to retrieve.}

\item{\code{...}}{Additional parameters as required by the OpenAI API.}

\item{\code{verbosity}}{numeric Verbosity level for the API call(0:no output;1:show headers;
2:show headers and bodies;3: show headers, bodies, and curl status messages.)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The modified run object matching the specified ID.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-openai-runs_list"></a>}}
\if{latex}{\out{\hypertarget{method-openai-runs_list}{}}}
\subsection{Method \code{runs_list()}}{
Returns a list of runs for a given thread.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{openai$runs_list(thread_id, ..., verbosity = 0)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{thread_id}}{character Required The ID of the thread the run belongs to.}

\item{\code{...}}{Additional parameters as required by the OpenAI API.}

\item{\code{verbosity}}{numeric Verbosity level for the API call(0:no output;1:show headers;
2:show headers and bodies;3: show headers, bodies, and curl status messages.)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list of run objects.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-openai-runs_submit_tool_outputs"></a>}}
\if{latex}{\out{\hypertarget{method-openai-runs_submit_tool_outputs}{}}}
\subsection{Method \code{runs_submit_tool_outputs()}}{
When a run has the status: "requires_action" and required_action.type is submit_tool_outputs,
             this endpoint can be used to submit the outputs from the tool calls once they're all completed.
             All outputs must be submitted in a single request.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{openai$runs_submit_tool_outputs(thread_id, run_id, tool_outputs, verbosity = 0)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{thread_id}}{character Required The ID of the thread the run belongs to.}

\item{\code{run_id}}{character Required The ID of the run to retrieve.}

\item{\code{tool_outputs}}{character Required. A list of tools for which the outputs are being submitted.}

\item{\code{verbosity}}{numeric Verbosity level for the API call(0:no output;1:show headers;
2:show headers and bodies;3: show headers, bodies, and curl status messages.)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The modified run object matching the specified ID.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-openai-runs_cancel"></a>}}
\if{latex}{\out{\hypertarget{method-openai-runs_cancel}{}}}
\subsection{Method \code{runs_cancel()}}{
Cancels a run that is in_progress.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{openai$runs_cancel(thread_id, run_id, verbosity = 0)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{thread_id}}{character Required The ID of the thread the run belongs to.}

\item{\code{run_id}}{character Required The ID of the run to retrieve.}

\item{\code{verbosity}}{numeric Verbosity level for the API call(0:no output;1:show headers;
2:show headers and bodies;3: show headers, bodies, and curl status messages.)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The modified run object matching the specified ID.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-openai-runs_create_tread"></a>}}
\if{latex}{\out{\hypertarget{method-openai-runs_create_tread}{}}}
\subsection{Method \code{runs_create_tread()}}{
Create a thread and run it in one request.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{openai$runs_create_tread(assistant_id, ..., verbosity = 0)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{assistant_id}}{character Required The ID of the assistant to use to execute this run.}

\item{\code{...}}{Additional parameters as required by the OpenAI API.}

\item{\code{verbosity}}{numeric Verbosity level for the API call(0:no output;1:show headers;
2:show headers and bodies;3: show headers, bodies, and curl status messages.)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A run object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-openai-runs_steps_retrieve"></a>}}
\if{latex}{\out{\hypertarget{method-openai-runs_steps_retrieve}{}}}
\subsection{Method \code{runs_steps_retrieve()}}{
Retrieves a run step.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{openai$runs_steps_retrieve(thread_id, run_id, step_id, verbosity = 0)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{thread_id}}{character Required. The ID of the thread to which the run and run step belongs.}

\item{\code{run_id}}{character Required. The ID of the run the step belongs to.}

\item{\code{step_id}}{character Required. The ID of the step to retrieve.}

\item{\code{verbosity}}{numeric. Verbosity level for the API call(0:no output;1:show headers;
2:show headers and bodies;3: show headers, bodies, and curl status messages.)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The run step object matching the specified ID.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-openai-runs_steps_list"></a>}}
\if{latex}{\out{\hypertarget{method-openai-runs_steps_list}{}}}
\subsection{Method \code{runs_steps_list()}}{
Returns a list of run steps belonging to a run.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{openai$runs_steps_list(thread_id, run_id, ..., verbosity = 0)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{thread_id}}{character Required The ID of the thread the run belongs to.}

\item{\code{run_id}}{character Required The ID of the run the step belongs to.}

\item{\code{...}}{Additional parameters as required by the OpenAI API.}

\item{\code{verbosity}}{numeric Verbosity level for the API call(0:no output;1:show headers;
2:show headers and bodies;3: show headers, bodies, and curl status messages.)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list of run step objects.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-openai-clone"></a>}}
\if{latex}{\out{\hypertarget{method-openai-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{openai$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
