% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threads.R
\name{threads}
\alias{threads}
\title{threads Class}
\description{
threads Class

threads Class
}
\section{Super class}{
\code{\link[openaistream:openai]{openaistream::openai}} -> \code{threads}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-threads-create}{\code{threads$create()}}
\item \href{#method-threads-retrieve}{\code{threads$retrieve()}}
\item \href{#method-threads-modify}{\code{threads$modify()}}
\item \href{#method-threads-delete}{\code{threads$delete()}}
\item \href{#method-threads-clone}{\code{threads$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="openaistream" data-topic="openai" data-id="initialize"><a href='../../openaistream/html/openai.html#method-openai-initialize'><code>openaistream::openai$initialize()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-threads-create"></a>}}
\if{latex}{\out{\hypertarget{method-threads-create}{}}}
\subsection{Method \code{create()}}{
Create a thread.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{threads$create(..., verbosity = 0)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Additional parameters as required by the OpenAI API.For example:messages;name;metadata}

\item{\code{verbosity}}{numeric Verbosity level for the API call(0:no output;1:show headers;
2:show headers and bodies;3: show headers, bodies, and curl status messages.)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A thread object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-threads-retrieve"></a>}}
\if{latex}{\out{\hypertarget{method-threads-retrieve}{}}}
\subsection{Method \code{retrieve()}}{
Retrieves a thread.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{threads$retrieve(thread_id, verbosity = 0)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{thread_id}}{character Required. The ID of the thread to retrieve.}

\item{\code{verbosity}}{numeric Verbosity level for the API call(0:no output;1:show headers;
2:show headers and bodies;3: show headers, bodies, and curl status messages.)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The thread object matching the specified ID.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-threads-modify"></a>}}
\if{latex}{\out{\hypertarget{method-threads-modify}{}}}
\subsection{Method \code{modify()}}{
Modifies a thread.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{threads$modify(thread_id, ..., verbosity = 0)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{thread_id}}{The ID of the thread to modify. Only the metadata can be modified.}

\item{\code{...}}{Additional parameters as required by the OpenAI API.For example:metadata}

\item{\code{verbosity}}{numeric Verbosity level for the API call(0:no output;1:show headers;
2:show headers and bodies;3: show headers, bodies, and curl status messages.)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The modified thread object matching the specified ID.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-threads-delete"></a>}}
\if{latex}{\out{\hypertarget{method-threads-delete}{}}}
\subsection{Method \code{delete()}}{
Delete a thread.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{threads$delete(thread_id, verbosity = 0)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{thread_id}}{character Required The ID of the thread to delete.}

\item{\code{verbosity}}{numeric Verbosity level for the API call(0:no output;1:show headers;
2:show headers and bodies;3: show headers, bodies, and curl status messages.)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Deletion status.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-threads-clone"></a>}}
\if{latex}{\out{\hypertarget{method-threads-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{threads$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
